Separated components into individual files
- Create container.spec.js
- Test that container renders child components
