+++ src/components/UrlFilter.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+export class UrlFilter extends React.Component {
+  render () {
+    return (
+      <div> Filter Placeholder </div>
+    );
+  }
+}
